refactor: extract jwt validation from `routes/task.js`
