+++ backend/middleware/auth.js
@@ -1,0 +1,17 @@
+import jwt from 'jsonwebtoken';
+
+export const verifyToken = (req, res, next) => {
+    const token = req.header('Authorization');
+    if (!token || !token.startsWith('Bearer ')) {
+        return res.status(401).json({ error: 'No token provided' });
+    }
+
+    try {
+        const decoded = jwt.verify(token.split(' ')[1], process.env.JWT_SECRET);
+        req.user = decoded;
+        next();
+    } catch (error){
+        console.error("JWT Error:", error.message);
+        res.status(401).json({ error: 'Access denied' });
+    }
+}
